Change unit types

+++ db/migrate/20230626101445_create_recipes.rb
@@ -2,8 +2,8 @@ class CreateRecipes < ActiveRecord::Migration[7.0]
   def change
     create_table :recipes do |t|
       t.string :name
-      t.integer :preparation_time
-      t.integer :cooking_time
+      t.decimal :preparation_time
+      t.decimal :cooking_time
       t.text :description
       t.boolean :public
       t.integer :user_id
